// packages/sigidep-client/src/app/utils/etat-mandat.enum.ts
export enum EtatMandatEnum {
  ANNULATIONMANDAT = 'labels.annulationduMandat',
  ANNULELORSRESERVATION = 'labels.annulelorsdelareservation',
  REJETLORSRESERVATION = 'labels.rejetlorsdelareservation',
  REJETCONTROLECONFORMITE = 'labels.rejetlorsduControleConformite',
  REJETCONTROLEREGULARITE = 'labels.rejetlorsduControledeRegularite',
  REJETVALIDATIONACT = 'labels.rejetlorsdelavalidationACT',
  MANDATMODIFIE = 'labels.mandatModifie',
  MANDATENREGISTRE = 'labels.mandatEnregistre',
  MANDATRESERVE = 'labels.mandatReserve',
  VISADAAF = 'labels.visaDAAF',
  TRANSMISCONTROLECONFORMITE = 'labels.transmispourControleConformite',
  RECEPTIONCONTROLECONFORMITE = 'labels.receptionpourControleConformite',
  CONTROLECONFORMITE = 'labels.controledeConformite',
  EDITIONTITRECREANCE = 'labels.editionduTitredeCreance',
  TRANSMISSIONLIQUIDATION = 'labels.transmissionpourLiquidation',
  RECEPTIONLIQUIDATION = 'labels.receptionpourLiquidation',
  SAISIEUNITESPHYSIQUESBIP = 'labels.saisiedesUnitesPhysiquesBIP',
  ENREGISTREMENTLIQUIDATION = 'labels.enregistrementdelaliquidation',
  LIQUIDATIONMODIFIEE = 'labels.liquidationmodifiee',
  VALIDATIONLIQUIDATION = 'labels.validationdelaLiquidation',
  ORDONNANCEMENT = 'labels.ordonnancement',
  TRANSMISSIONCONTROLEDEREGULARITE = 'labels.transmissionpourControledeRegularite',
  RECEPTIONCONTROLEREGULARITE = 'labels.receptionpourControleRegularite',
  CONTROLEREGULARITE = 'labels.controledeRegularite',
  TRANSMISSIONACT = 'labels.transmissionACT',
  RECEPTIONACT = 'labels.receptionACT',
  VALIDATIONCOMPTABLE = 'labels.validationComptable',
  PAIEMENT = 'labels.paiement',
}